+++ PyrusServiceDeskAndroid/pyrusservicedesk/src/main/java/com/pyrus/pyrusservicedesk/sdk/data/AttachmentDto.kt
@@ -5,14 +5,14 @@ import com.google.gson.annotations.SerializedName
 import com.pyrus.pyrusservicedesk.presentation.ui.view.Status
 import com.pyrus.pyrusservicedesk.sdk.data.gson.Local
 
-private const val FILE_ID_EMPTY = 0
+private const val FILE_ID_EMPTY = 0L
 
 /**
  * Represents an attachment of the comment.
  * @param localUri transient field that is used only for local comments.
  */
 internal data class AttachmentDto(
-    @SerializedName("id") val id: Int = FILE_ID_EMPTY,
+    @SerializedName("id") val id: Long = FILE_ID_EMPTY,
     @SerializedName("guid") val guid: String = "",
     @SerializedName("type") val type: Int = 0,
     @SerializedName("name") val name: String = "",
